Split out common and builtin linalg
