+++ CMakeLists.txt
@@ -68,6 +68,7 @@ endif()
 
 set_property(GLOBAL APPEND PROPERTY QOCO_SOURCES "${CMAKE_CURRENT_SOURCE_DIR}/src/qoco_api.c"
     ${CMAKE_CURRENT_SOURCE_DIR}/src/input_validation.c
+    ${CMAKE_CURRENT_SOURCE_DIR}/src/common_linalg.c
     ${CMAKE_CURRENT_SOURCE_DIR}/src/kkt.c
     ${CMAKE_CURRENT_SOURCE_DIR}/src/cone.c
     ${CMAKE_CURRENT_SOURCE_DIR}/src/qoco_status.c
